Use LCP Client in Tracer (#5908)
Move the LCP calls to the client.
