+++ pyproject.toml
@@ -105,7 +105,7 @@ singlestoredb = {version = "^0.6.1", optional = true}
 pyspark = {version = "^3.4.0", optional = true}
 tigrisdb = {version = "^1.0.0b6", optional = true}
 nebula3-python = {version = "^3.4.0", optional = true}
-langchainplus-sdk = ">=0.0.6"
+langchainplus-sdk = ">=0.0.7"
 
 
 [tool.poetry.group.docs.dependencies]
